requirements.txt
anytree==2.6.0
apparmor==2.13.3
appdirs==1.4.3
asn1crypto==0.24.0
astroid==2.0.4
backcall==0.1.0
Beaker==1.10.1
blinker==1.4
brotlipy==0.7.0
btrfsutil==1.1.1
CacheControl==0.12.5
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
colorama==0.4.1
cryptography==2.3.1
decorator==4.4.0
distlib==0.2.9
distro==1.4.0
dlib==19.17.0
docutils==0.14
face-recognition==1.2.3
face-recognition-models==0.3.0
future==0.17.1
GDAL==3.0.0
h11==0.7.0
h2==3.1.0
hpack==3.0.0
html5lib==1.0.1
hyperframe==5.2.0
idna==2.8
iotop==0.6
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.1.0
isc==2.0
isort==4.3.4
jedi==0.13.3
jupyter-client==5.2.4
jupyter-core==4.5.0
kaitaistruct==0.8
lazy-object-proxy==1.3.1
ldap3==2.5.2
lensfun==0.3.2
LibAppArmor==2.13.3
lockfile==0.12.2
louis==3.10.0
lxml==4.3.4
Mako==1.0.12
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
meson==0.51.0
msgpack==0.6.1
numpy==1.16.4
onboard==1.4.1
packaging==19.0
parso==0.5.0
passlib==1.7.1
pep517==0.5.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
progress==1.5
progressbar2==3.42.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pwquality==1.4.0
pyasn1==0.4.5
pycairo==1.18.1
pycparser==2.19
Pygments==2.4.2
PyGObject==3.32.2
pylint==2.1.1
pyOpenSSL==18.0.0
pyparsing==2.2.2
pyperclip==1.6.5
python-dateutil==2.8.0
python-utils==2.3.0
pytoml==0.1.20
pyzmq==18.0.1
Reflector==2019.3.8.1.54.39
requests==2.22.0
retrying==1.3.3
ruamel.yaml==0.15.89
simplegeneric==0.8.1
six==1.12.0
sortedcontainers==2.0.5
speedtest-cli==2.1.1
tabulate==0.8.3
team==1.0
tetherback==0.9.1
tornado==5.1.1
tqdm==4.32.2
traitlets==4.3.2
urllib3==1.25.3
urwid==2.0.1
virtualenv==16.0.0
wcwidth==0.1.7
webencodings==0.5.1
wrapt==1.10.11
wsproto==0.11.0
youtube-dl==2019.6.27
